DKAM: Add ubuntu OS image download functionality. (#22)
Add functionality to download jammy server image.
Convert .img file to raw.gz
Place jammy-server-cloudimg-amd64.raw.gz file in PVC.
Update proto file with responce struct.

+++ mi-dkam/pkg/script/secure_hookos.sh
@@ -99,6 +99,9 @@ sign_all_components() {
         mv /data/hook_x86_64.tar.gz /data/unsigned
     else
         echo "Path /data does not exist."
+        if [ ! -f $working_dir/hook_x86_64.tar.gz ]; then
+            exit 0
+        fi
         tar -xvf $working_dir/hook_x86_64.tar.gz -C $STORE_ALPINE
         mkdir -p $working_dir/unsigned
         mv $working_dir/hook_x86_64.tar.gz $working_dir/unsigned
